@@ -1,22 +1,19 @@
-import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot, UrlTree } from '@angular/router';
-import { Observable } from 'rxjs';
+import { inject } from '@angular/core';
+import { CanActivateFn } from '@angular/router';
+import { AutenticacaoService } from '../auth/autenticacao.service';
 
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthGuardsGuard implements CanActivate {
-  canActivate(
-    route: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+export const authGuardsGuard: CanActivateFn = (route, state) => {
+
+
     const authService = inject(AutenticacaoService);
     const user = authService.getRoles();
     const acesso = route.data['roles'];
 
-    if(user.acesso.includes(acesso)){
+    if(user.roles.includes(acesso)){
       return true;
     }
+
     return false;
-  }
-  
-}
+    
+    
+ };
