@@ -1,10 +1,10 @@
 import { NgModule} from '@angular/core';
 import { RouterModule, Routes} from '@angular/router';
 import { HomeComponent} from './home/home.component';
-import { AuthGuardsGuard } from '../guards/auth-guards.guard';
+import { authGuardsGuard } from '../guards/auth-guards.guard';
 
 const routes: Routes = [{
-    path: '', component: HomeComponent, canActivate:[AuthGuardsGuard], data:{ roles:'ROLE_USER'}
+    path: '', component: HomeComponent, canActivate:[authGuardsGuard], data:{ roles:'ROLE_USER'}
 }];
 
 @NgModule({
